Fix #10: mozilla/mozilla/mozilla/mozilla

+++ mozilla/.profile
@@ -2,9 +2,13 @@ xdg_moz_dir="$XDG_CONFIG_HOME/mozilla"
 
 if [ "$(uname -s)" = Darwin ]; then
     grr_dir="$HOME/Library/Application Support/Firefox"
-    rm -r "$grr_dir"
-    ln -sf "$xdg_moz_dir/firefox/" "$grr_dir"
+    if ! [ -L "$grr_dir" ]; then
+        rm -r "$grr_dir"
+        ln -s "$xdg_moz_dir/firefox/" "$grr_dir"
+    fi
 else
-    rm -r "$HOME/.mozilla"
-    ln -sf "$xdg_moz_dir" "$HOME/.mozilla"
+    if ! [ -L "$HOME/.mozilla" ]; then
+        rm -r "$HOME/.mozilla"
+        ln -s "$xdg_moz_dir" "$HOME/.mozilla"
+    fi
 fi
